refactor: restructure code into packages
