+++ pkg/scanners/scan_crowdstrike.go
@@ -1,4 +1,6 @@
-package edrRecon
+package scanners
+
+import "github.com/FourCoreLabs/EDRHunt/pkg/resources"
 
 type CrowdstrikeDetection struct{}
 
@@ -6,8 +8,8 @@ func (w *CrowdstrikeDetection) Name() string {
 	return "Crowdstrike EDR Solution"
 }
 
-func (w *CrowdstrikeDetection) Type() EDRType {
-	return SymantecEDR
+func (w *CrowdstrikeDetection) Type() resources.EDRType {
+	return resources.SymantecEDR
 }
 
 var CrowdstrikeHeuristic = []string{
@@ -16,11 +18,11 @@ var CrowdstrikeHeuristic = []string{
 	"%SYSTEMROOT%\\system32\\drivers\\crowdstrike\\CsFirmwareAnalysis.inf",
 }
 
-func (w *CrowdstrikeDetection) Detect(data SystemData) (EDRType, bool) {
+func (w *CrowdstrikeDetection) Detect(data resources.SystemData) (resources.EDRType, bool) {
 	_, ok := data.CountMatchesAll(CrowdstrikeHeuristic)
 	if !ok {
 		return "", false
 	}
 
-	return CrowdstrikeEDR, true
+	return resources.CrowdstrikeEDR, true
 }
