+++ pkg/scanners/scan_sentinelone.go
@@ -1,0 +1,28 @@
+package scanners
+
+import "github.com/FourCoreLabs/EDRHunt/pkg/resources"
+
+type SentinelOneDetection struct{}
+
+func (w *SentinelOneDetection) Name() string {
+	return "SentinelOne"
+}
+
+func (w *SentinelOneDetection) Type() resources.EDRType {
+	return resources.SentinelOneEDR
+}
+
+var SentinelOneHeuristic = []string{
+	"SentinelOne\\",
+	"CbDefense\\",
+	"SensorVersion",
+}
+
+func (w *SentinelOneDetection) Detect(data resources.SystemData) (resources.EDRType, bool) {
+	_, ok := data.CountMatchesAll(SentinelOneHeuristic)
+	if !ok {
+		return "", false
+	}
+
+	return resources.SentinelOneEDR, true
+}
