+++ pkg/scanners/scanner.go
@@ -1,0 +1,17 @@
+package scanners
+
+import "github.com/FourCoreLabs/EDRHunt/pkg/resources"
+
+var (
+	Scanners = []resources.EDRDetection{
+		&CarbonBlackDetection{},
+		&CrowdstrikeDetection{},
+		&CylanceDetection{},
+		&FireEyeDetection{},
+		&KaskperskyDetection{},
+		&McafeeDetection{},
+		&SymantecDetection{},
+		&SentinelOneDetection{},
+		&WinDefenderDetection{},
+	}
+)
