+++ pkg/util/util.go
@@ -1,4 +1,4 @@
-package edrRecon
+package util
 
 import "strings"
